requirements.txt
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.1.0
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
passlib==1.7.4
protobuf==4.21.12
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
requests==2.29.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
starlette==0.26.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
webdriver-manager==4.0.0
wsproto==1.2.0
boto3
